major cleanup, very close now
